Добавлена работа с ресурсами с помощью примитивного потока int  и фильтра
