+++ testlambdas/resources/data.txt
@@ -1,0 +1,1 @@
+52, 654, 378, -977, 87, 873, 777, 2987564, 9
