+++ testlambdas/src/testapp/AppStream.java
@@ -1,3 +1,5 @@
+package testapp;
+
 import java.util.ArrayList;
 import java.util.List;
